Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

Common script == awesome idea.

I'm going to have to agree with Sean. At least half of the file (if not more) is workarounds for lsb_release possibly not being there. The rest is adding an APT repo and then installing a few packages (and in one case, maybe installing apparmor files, which I gotta say I'm confused aren't just in the docker-engine deb, but whatev)

Unpacking that into the lib/docker script directly - without all of the logic for working around missing lsb_release - and then adding an infra-hosted mirror of the docker repo with the ability to turn it on like we do for ceph and UCA repos when jobs want to use them is going to be much healthier in the long run.

Patch-set: 1
Reviewer: Gerrit User 2 <2@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1