Update patch set 2
Patch Set 2: Code-Review+2

we used have this "literate programming" thing (shocco) that generated a sort of verbose document from the script directly.  i don't know if it supported rst comments but it look like it came from that.  or it was copy pasted.

I'm not much for code-motion but I guess let's err on the side of less confusion.  committing as doc only.

Patch-set: 2
Label: Code-Review=+2