Update patch set 11
Patch Set 11:

> interestingly, placement seems totally broken in this patch, though
 > I'm not super clear why

I suspect: With keystone under uwsgi instead of mod-wsgi, the <Directory /usr/bin> stuff that keystone under mod-wsgi adds but placement depends on is no longer present in the global space of the apache configurtion. So one option, if we want to keep the patches separate is to add those lines into placement's apache file.

Patch-set: 11