Update patch set 3
Patch Set 3:

The grenade fails are expected, we're going to have to pin the wsgi mode for those. 

The up/down is because "keystone" is no longer an apache site to be disabled (there are now 2 of them), we'll need to select on that.

Centos 7 is not our fault, that's the epel remove tripping up cinder.

The multinode fail looks super early and spurious (node killed before devstack started).

The swift functional tests are because they are hard coding keystone's address (and not pulling it out of the openrc)

Patch-set: 3
Reviewer: Gerrit User 2750 <2750@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0