Update patch set 1
Patch Set 1:

> The commit title is not descriptive enough

IMO, the commit title + commit message is more than adequate for a new contributor patch that fixes the spelling of a single word in a comment.

It's hard enough to get OpenStack reviews as it is, and a -1 on a review often  means a core won't look at it now.

In the future, I would consider using a zero vote for such cases, and also perhaps suggesting the commit title you would have preferred to see. Perhaps:

- "Fix typo: retreive -> retrieve"

Patch-set: 1