Update patch set 1
Patch Set 1:

> It's hard enough to get OpenStack reviews as it is, and a -1 on a review often  means a core won't look at it now.

Agreed, but as a core reviewer is going to -1 for the same reason, let's save some time here.

In addition to what you a fixing ("typo in a code comment"), an indication as to where/what is going to be fixed is helpful: "Fix typo in code comment of tools/dstat.sh"

Stating you've changed retreive into retrieve is not top priority in the commit title, this doesn"t give an hint on "who is best suited to review this".

Patch-set: 1