Update patch set 1
Patch Set 1:

That 50 char limitation is invalid to me. 

I've been involved in the community for several years, I guess people are easier on me. But when you start, you have to pay extra attention to details and rules.

Anyway, I am not sure what's the big deal here. Changing that title takes 5sec and I -1'ed that patch because, to me, at that time, it's not good enough to be merged.

Patch-set: 1