[WIP] Use pbr installed nova-wsgi-api script
This is a precursor to getting nova running as a wsgi app under any
container. This is checking that it will work in the old way.

The depended-on code provides a wsgi application that:

* is installed by pbr via an entry point defined in setup.cfg
* properly handles setting the service version for osapi_computer
* properly fails when an out of date version is run

This will be intergrated with the ongoing systemd/uwsgi/portless
work that is in progress once the edges have been ground off.

Change-Id: Ia7d8eee2a1a9ca08f829962cad19ed638629828b
Depends-On: I7c4acfaa6c50ac0e4d6de69eb62ec5bbad72ff85