Update patch set 1
Patch Set 1: Workflow-1

this isn't working because tls proxy is on in the tests, expecting an internal port for nova-api, which is not set up when nova-api is in wsgi mode.

So I suspect the right thing to do here is to make a new change based off the top of the uwsgi changes and clean up more significantly.

Patch-set: 1
Reviewer: Gerrit User 11564 <11564@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Workflow=-1