[WIP] Use uwsgi for nova-api
This is a hack job to experiment with nova-api and nova-api-meta
under uwsgi.

The depends-on is using the pbr installed wsgi scripts that
are aware of service versions.

The change is functions-common is for the case where a service which
is not going to be under systemd had not been before so there are no
unit files to stop. Without the /bin/true the lines can fail during
nova_cleanup. This could/should be moved to another patch once this
is proven.

Because the metadata server is always expected to run on a port and
without a prefix, we have it configured to use uwsgi but not to
proxy from apache: uwsgi listens on the configured port itself.

uwsgi process taht listen themselve do not need a socket or to
chmod-socket, so those config lines have been moved to the block
that is also writing proxy configuration for apache.

Because this change only uses uwsgi for nova-api and nova-api-meta,
nova-api-meta is set to default to enabled in stackrc because the
nova-api wsgi application used by wsgi only presents the one service
(osapi_compute).

Change-Id: If2d7e363a6541854f2e30c03171bef7a41aff745
Depends-On: Icb35fe2b94ab02c0ba8ba8129ae18aae0f794756