Use uwsgi for nova when NOVA_USE_MOD_WSGI is not False
Unless NOVA_USE_MOD_WSGI is False, run nova-api and nova-metadata
using uwsgi.

The change in functions-common is for the case where a service which
is not going to be under systemd had not been before so there are no
unit files to stop. Without the /bin/true the lines can fail during
nova_cleanup. This could/should be moved to another patch once this
is proven.

Because the metadata server is always expected to run on a port and
without a prefix, we have it configured to use uwsgi but not to
proxy from apache: uwsgi listens on the configured port itself.

uwsgi process that listen themselve do not need a socket or to
chmod-socket, so those config lines have been moved to the block
that is also writing proxy configuration for apache.

Because this change only uses uwsgi for nova-api and nova-api-meta,
nova-api-meta is set to default to enabled in stackrc because the
nova-api wsgi application used by wsgi only presents the one service
(osapi_compute).

Change-Id: If2d7e363a6541854f2e30c03171bef7a41aff745
Depends-On: Icb35fe2b94ab02c0ba8ba8129ae18aae0f794756
Depends-On: I7b22692ea9135547d93b08724f1f9ef060f13007