[WIP] Use uwsgi for nova-api
This is a hack job to experiment with nova-api and nova-api-meta
under uwsgi.

It removes all the mod-wsgi stuff, for the sake of clean slate. It
ignores running the metadata service at all, for the time being.

The depends-on is using the pbr installed wsgi scripts that
are aware of service versions.

The change is functions-common is for the case where a service which
is not going to be under systemd had not been before so there are no
unit files to stop. Without the /bin/true the lines can fail during
nova_cleanup.

Change-Id: If2d7e363a6541854f2e30c03171bef7a41aff745
Depends-On: Icb35fe2b94ab02c0ba8ba8129ae18aae0f794756