[WIP] Use uwsgi for nova-api
This is a hack job to experiment with nova-api under uwsgi.

It removes all the mod-wsgi stuff, for the sake of clean slate. It
ignores running the metadata service at all, for the time being.

The depends-on is using the pbr installed nova-wsgi-api script that
is aware of service versions.

Change-Id: If2d7e363a6541854f2e30c03171bef7a41aff745
Depends-On: I7c4acfaa6c50ac0e4d6de69eb62ec5bbad72ff85