Update patch set 1
Patch Set 1:

What's the usage here?

Nova's got a noauth middleware as well, but it really feels like we should instead make a keystone permissive mode instead of changing the auth bit, which is going to impact the ability to do roles and limits in the future.

Patch-set: 1
Label: Verified=0