WIP Replace LC_ALL=C with LC_ALL=en_US.UTF-8
This replaces the remaining LC_ALL=C with LC_ALL=en_US.UTF-8 to get
unicode support and get consistent with the other settings of LC_ALL in
devstack. We would try to use LC_ALL=C.UTF-8 but it isn't included in
centos 7 out-of-the-box.

We'll also export PYTHONIOENCODING=UTF-8 to cover all cases where we
run python.

Related-Bug: #1580728

Change-Id: I6f36b359f4623ca6d74eb0ce0e338ca64c432e04