Restore log colors under devstack/systemd
One of the pending issues with the conversion to systemd was the loss of
log coloring.  It turns out that journalctl by default strips out
characters it considers "unprintable" - including the color codes
emitted by the old-style logging.  However, journalctl can be made to
print them by adding the `-a` flag.

This change makes devstack's log formatter conf settings include color
codes like the old screen-based setup used to; and adds the `-a` flag to
the journalctl command that copies the logs into regular text files.

Change-Id: I2401e267913a24d18dae355aa933072dbbdab1d8