Update patch set 26
Patch Set 26: Code-Review-1

This patch effectively breaks devstack on any arch that is not in the list, e.g. ppc64. The commit message also does not explain why it is needed as a base service. Was there a discussion that everyone knows about, but which is not linked in commit?

Do I now need to remove etcd from the list of base services if I need to use devstack on a non-listed arch?

Patch-set: 26
Label: Code-Review=-1