Update patch set 27
Patch Set 27:

Mikhail,

Added some defensive code to detect if a package is present and just restart the service if it is present. This should enable you to install etcd3 however you wish before starting devstack.

Note that we are using etcd3 3.1.x from https://github.com/coreos/etcd/releases/ It does not currently have options to install on ppc64. 3.2.0 is in Release candidate mode, so we can switch to that when it gets available, hopefully not too far away. That has more options to install and we can start doing that.

Thanks,
Dims

Patch-set: 27
Label: Verified=0