Make cinder service enablement clearer
This makes cinder service management more direct and clear. If the
"cinder" alias is explicitly disabled then don't report it as enabled
even if sub service may be enabled. This allows for simpler overrides of
the cinder service being enabled by default.

Also only run the cinder api if the cinder api is enabled. This includes
checking for whether or not the service came up.

Change-Id: I1c4f22f785af96caaf4baa21ff28714b9afd3458