Update patch set 1
Patch Set 1:

I actually care more that it's relative to DATA_DIR than that it be configurable.  All other data is stored under DATA_DIR and I believe this should also be for consistency.  I have a DATA_DIR that is in a separate directory (not under DEST), and when I re-initialize my devstack environment I always clean out my DATA_DIR.  With this hardcoded, when I re-initialized my devstack this was left over and caused failures. I can manually clean this out as well, but it feels like it should fit into the same pattern as other data.  With the default definition of DATA_DIR, this ps doesn't actually change the ETCD_DATA_DIR location - just the way it's specified.

Making it configurable is secondary, IMO.  In the devstack architecture, it appears that almost all variables are defined with the FOO=${FOO:-defaultvalue} syntax which allows them to be configured by the user.  Therefore, having this be configurable just seems more consistent with the rest of the devstack design.

I'm certainly willing to back out making it customizable and just specify it as ${DATA_DIR}/etcd, if that is the preference.  This would meet my needs and would at least put it in the same location as other devstack data.

Patch-set: 1
Reviewer: Gerrit User 24147 <24147@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0