Make stack.sh work on SUSE
This adds packages to suse for systemd python linkages as well as
apache2 and which. And configures mod_proxy and mod_proxy_uwsgi with
a2enmod.

We also properly query if apache mods are enabled to avoid running
into systemd service restart limits.

Also in addition to enabling mod_ssl we enable the SSL flag so that TLS
will work...

Change-Id: I3442cebfb2e7c2550733eb95a12fab42e1229ce7