Run devstack CA and cert setup early
Previously apache was configured and restarted before we configured
the CA and certs. In most cases this is fine because those specific
vhosts didn't use tls. However, if you had previously run devstack
and had leftover vhosts and an unconfigured CA or certs devstack would
fail.

This is a small corner case, but its simple to address by moving CA and
cert setup up in stack.sh to before we do anything related to web
servers.

We also trust the devstack CA early so that services trust it when
started, easy mode for making sure TLS works everywhere (without this
you have to explicitly configure each service to trust the CA anyways).

Change-Id: I31dbaf9471088b9faff26c7b790da6f6feebb2d5