Set the cafile in swift test.conf if tls enabled
Because the swift functests (which use test.conf) run out of a
virtualenv they don't get access to the system wide trust of the
devstack CA. Handle this by explicitly configuring the cafile to trust
in the test.conf file.

Swift's functests will need to read this in and properly configure
things on its end.

Change-Id: I4cdba36ccab6acd76205184882ee29e4f1e12333