Add option to not install test-requirements.txt
The more software we install the longer it takes to run devstack and the
slower tools like OSC run. Avoid installing test-requirements to reduce
the total python package installation size.

Note that this currently defaults to not installing the test
requirements but rough plan would be to test it this way, update d-g to
set this var to False, then switch default to True in this change and
merge it. This way devstack itself is backward compatible but the test
runs by zuul should be sped up.

Depends-On: Ic5f1844ce732d447ee980b3c9fdc417f72482609
Change-Id: I2fd639b9005b3a23a1bc37b8424da8fc100c48f6