Update patch set 3
Patch Set 3:

@ian,

This time I'm stepping up and saying reducing the complexity of the hash by default everywhere is incorrect. The reason this is lowered in devstack is because the resources for devstack are more constrained than a real production system.

We will look at tuning the defaults in keystone against a reasonable set of deployment settings (it may not result in a default change at all). Bcrypt will not be tuned for devstack.

It is *VERY* important to ensure that the defaults of openstack maintain a high-level of integrity for the stored password hashes and reducing the rounds explicitly sacrifices time-complexity in production (which should be more secure) for an under-powered development/CI-allin-in-one-node.

The fix is being pushed to devstack, acknowledging we will at least need to provide solid guidance on the  configuration and tuning of keystone. Auth needs to have the high time-complexity (or in the case of scrypt) time-memory-complexity to ensure the hashes are not harder to brute-force.

In short, the previous fix should not have been upstream. This one absolutely should not have be upstream. Keystone should default to more secure and allow operators to opt to trade faster auth for lesser security.

Patch-set: 3
Reviewer: Gerrit User 2903 <2903@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0