Update patch set 1
Patch Set 1:

heh, is devstack a test or not?  everything old is new again :)

I've just been around long enough to remember the "test-only" packages variable we had ... both when we started installing python deps [1] and decided it was default behaviour to do this [2].

TBH I think that people do non-production stuff in devstack that expects these to be here, and was how we ended up with this, so installing them is just part of the price of running devstack.  

Maybe, however, we have traded off too much towards convenience over speed by doing this, and we should do revert?

[1] https://review.openstack.org/#/c/107001
[2] https://review.openstack.org/#/c/167669

Patch-set: 1
Label: Verified=0