Update patch set 1
Patch Set 1:

There are two related issues here. The first is that pip installation time just keeps getting longer and longer. Pip isn't quick, even when using wheels. This appears in part to be related to needing to evaluate all the inputs in order to do an install (existing packages, constraints, env markers, requirements, so on). General network and disk cost being the other significant slowdown here.

The second is that importing package resources lists and sorts all installed packages which directly affects the runtime of tools like openstackclient which devstack makes extensive use of. We end up with every additional python package installed slowing everything down.

Comparing the pip listing for gate-tempest-dsvm-neutron-full-ubuntu-xenial on this change and its parent the difference is 61 packages (or about 20% of total package count) so its not tiny.

Patch-set: 1
Label: Verified=0