Update patch set 1
Patch Set 1:

Also worth noting that test-requirements isn't necessary to run the services (see jobs passing below). They should only be necessary to run functional tests and unittests which run out of virtualenvs that explicitly install the test-requirements if they need them.

I think this is largely overkill for convenience. Happy to flip this around so that we keep this behavior by default then flip the bit in devstack-gate for job runs.

Patch-set: 1