Install OSC into a virtualenv
The total number of packages you have installed affects the runtime of
openstackclient because it uses pkg_resources to find plugins and the
more python packages you have the slower finding plugins is. This is
likely to break any devstack plugin that uses OSC plugins because they
won't install to the correct openstackclient location. But we want to
test if this produces a measurable improvement to devstack runtimes then
worry about backward compat in a future patchset.

Change-Id: Iea80ae97254a70d15aeb7e586acc1bf175dc6a3a
Depends-On: Ic5f1844ce732d447ee980b3c9fdc417f72482609