Run OSC flavor create in parallel
Each invocation of OSC takes about 3-4 seconds. We run 10 of them just
to create nova's flavors. That comes out to about 30 seconds if we are
lucky. Or we can run them in parallel on 8 cpus and spend ~6 seconds
making flavors.

Use xargs to run things in parallel since it is a quick and easy way to
do so when you don't care about the process output (it will still exit
with an error if any subcommand errored). Note that this won't work with
the timing capturing code that sdague wrote for this reason.

Change-Id: I19a865388bc791758a2585264d81d4e620dea2ff