Update patch set 5
Patch Set 5: Code-Review-1

(3 comments)

I'm not opposed to this, but it seems to leave a number of places in limbo?  marked inline

there's also swift which seems to restart the service, but is it actually using it?

could we go so far as to remove it from the default package list, and just install it if etcd isn't there?

Patch-set: 5
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1