Update patch set 2
Patch Set 2:

I'm running into an issue when using this patch.  Devstack seems to want to use etcd even on single-node systems and this patch allows etcd to be installed, but doesn't start it.  Here's a link to the exception I get in cinder-volume: https://pastebin.com/raw/2Behpb1v

Patch-set: 2
Label: Verified=0