Make Cinders lvm_type default to auto
We've had an lvm_type=thin for quite some time, and we've wanted to
make it Cinders default for a while, but there's no clean way to
migrate on upgrades.  So, we have a compromise, detect if the
deployment has volumes (including the thin pool).  If there are no
LVs or if the thin pool exists, use thin.  Otherwise, use the old
thick lvm-default.

You can of course still set this in your local.conf if you wish
using "CINDER_LVM_TYPE=default".

This change will set the default to "auto" in devstack/lib/cinder
and specify that it be added to the local.conf.

Change-Id: Ia6b99c5df6614fe9b73b5731cf234e3acbf57496