Update patch set 1
Patch Set 1: Code-Review+1

Yes, unfortunately, it broke Ironic. We have 2 ironic patches to fix [1] & [2], but the first patch fails cuz of this. It would be good if we can revert this, get the ironic patches merged, then do this change. (Sorry. Alternatively, we could merge both ironic patches into one but there are a lot of changes...)

[1] Introduce keystoneauth adapters for clients: https://review.openstack.org/#/c/469627/

[2] keystone auth for glance: https://review.openstack.org/#/c/467728/

Patch-set: 1
Label: Code-Review=+1