Update patch set 1
Patch Set 1:

That's a possibility.  The alternative is something like https://review.openstack.org/#/c/409907/ where we tweak the cleanup code to be aware that it's a shared directory.

For what it's worth, the expected usage of the "unstack.sh" and "clean.sh" commands do not seem to be documented very well on https://docs.openstack.org/developer/devstack/

Patch-set: 1