Re-add commit message to devstack-version
As a follow on to I4c269a7f3d63ee9a976e7c3636fc3e5e8dab9ae3, when the
"cat << EOF" writes out the devstack-version script it does quote
removal on the ${git_version} string.  This means in the file you have
unquoted apostropies if someone uses "it's" etc.

If we just save it (quoted) to a var, then let the sub-script expand
it as usual we avoid this, and it reads sensibly.

Change-Id: Ic7391dd087657c0daf74046e4a052c53f4eb6e1a