Update patch set 1
Patch Set 1: Code-Review-1

Doh, devstack-version prints the last git log entry from *devstack*, not whatever other project it's testing. So in the CI run for [1] it just printed [2]:

DevStack Version: pike - 80804865ab2812f3a0bdfae5c4f4d238f5fa617c Re-add commit message to devstack-version 2017-06-30 12:32:06 +1000

which is the devstack version of *this patch*.

Given how badly this can break the gate, and that we've reverted the problematic change and are thus not in a hurry, I think I'd like to see this change temporarily modified to include special characters in the subject line. Once we're sure that passes CI, we can put back the regular subject line and merge.

[1] https://review.openstack.org/#/c/479127/
[2] http://logs.openstack.org/27/479127/3/check/gate-tempest-dsvm-py35-ubuntu-xenial/61db5ee/logs/devstacklog.txt.gz

Patch-set: 1
Label: Code-Review=-1