Update patch set 1
Patch Set 1: Code-Review-2

This is the wrong approach (ignoring the issues with the patch itself, which there are a couple as tosky started on) if a plugin doesn't properly expose it's requirements at install time it is broken. That's strictly a bug in the plugin and not something we should workaround here. This is a big motivation for why we're doing the community goal to split tempest plugins into separate python packages. Because the whole requirements space is a huge problem

Patch-set: 1
Reviewer: Gerrit User 5196 <5196@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-2