Enable tempest testing for extending an attached volume
In Pike, Cinder and Nova will support extending the size of
an attached volume, but it's not supported by all volume and
compute backends. Tempest will also test it but uses a config
option that we need to set based on how devstack is configured.

Depends-On: Ibace6c2f91be9753a44e5f79fd013df11654851b

Related to cinder blueprint extend-attached-volume
Related to nova blueprint nova-support-attached-volume-extend

Change-Id: I52cc2952a2938ce44c442aa3e3b69a905b2b55d5