Update patch set 3
Patch Set 3: Code-Review-1

Something blew up in the non-voting neutron multinode job which looks really odd:

http://logs.openstack.org/56/477556/3/check/gate-tempest-dsvm-neutron-multinode-full-ubuntu-xenial-nv/ee3e9b6/logs/screen-n-cond-cell1.txt.gz#_Jul_05_20_44_20_132760

This is for the test_create_server_with_scheduler_hint_group test, which is probably doing affinity testing, and the affinity lookup from the compute is disabled. Although this looks like a retry in the cell conductor that's failing.

Patch-set: 3
Label: Code-Review=-1