Update patch set 10
Patch Set 10:

So from a quick glance the code looks fine to me, but the commit message is a bit lacking I feel. There are some changes here that our outside just not using the ubuntu defaults and installing ldap-utils. For example, part of the fix here is to setup keystone with multiple domains per backend (which admittedly is a normalish way to configure things, but it's a pretty different way to do things than what was there before)

Also I'm wondering do we have an places where this is run, it's been about 6 or 7 yrs since I last deployed ldap, so I'd feel a bit more comfortable with a +2 if I could see this running somewhere.

Patch-set: 10
Label: Workflow=0