Fix image_list.sh sanity check
Patch [1] removed the trove image from stackrc. Now the image-list.sh
tools errors out with "We only found 4 images in [...]"

This also causes the DIB CI image build to fail, as due to the error
the fallback mechanism of determining the devstack images (element
cache_devstack) is used. The fallback is not able to deal with
variables. This leads to a repository list with entries like this:

  cirros-${CIRROS_VERSION}-${CIRROS_ARCH}-rootfs.img.gz

DIB then fails with "CIRROS_VERSION: unbound variable".

[1] https://github.com/openstack-dev/devstack/commit/729623de78688eb84902f7633e6c436a9a0874a6

Change-Id: If190393a1226f667eaa6efe5321c7f8340b98a38