Update patch set 1
Patch Set 1: Code-Review-1

this is more than a version bump, as it now ties it to the global variable.  I'm not sure if that's appropriate, @Bob?

anyway, assuming this is right, i'd like to see it called out in the changelog as explicitly something we want to be doing, for future code explorers.

Patch-set: 1
Label: Code-Review=-1