Update patch set 1
Patch Set 1: Code-Review-1

I don't understand the justification here.  The Cirros image for XenServer (note, this doesn't change the libvirt+xen case which is covered under libvirt) is downloaded from a Citrix website as the standard image can't be imported directly.
As such, we need to have separation from the global var - otherwise bumps to the libvirt cirros break XenServer until we can get a new image uploaded.  Generally that should be quite quick, but we don't want to block the libvirt cirros image change on not having a suitable image for XenServer.

Further, this is something that should actually be removed from devstack and moved to os-xenapi's devstack plugin.

Patch-set: 1
Label: Code-Review=-1