Tempest compute-feature-enabled overrides
This introduces two new variables, TEMPEST_COMPUTE_FEATURES_ENABLED
and TEMPEST_COMPUTE_FEATURES_DISABLED. These are lists of compute
features to enable or disable in the tempest configuration. If any
compute features are included in these lists, they will override any
previous setting of that feature in the tempest configuration.

Change-Id: I539bf351de6caf3f7393971d213bf4377a81fc7a