Update patch set 5
Patch Set 5:

For the experimental queue jobs, the nova-next failure was some random cinder related test, so not related.

The ceph py35 jobs are failing because of bug 1706405.

The opensuse jobs are failing because of some keystone multibackend configuration issue, so unrelated.

The only job failing that looks like it's caused by this is the neutron dvr-ha multinode job:

http://logs.openstack.org/56/477556/5/experimental/gate-tempest-dsvm-neutron-dvr-ha-multinode-full-ubuntu-xenial-nv/432c235/console.html#_2017-07-25_15_47_52_092840

Looking in logstash for that failure in that job, it's only happening on this change, so probably some issue in the setup there. Also note that's a 3-node job, so that's likely part of the issue (maybe we aren't accounting for the 3rd node in the setup here?). It sounds like haleyb in neutron cares about this job, and it was passing not too long ago, but it also seems to have gone off the rails since 7/21:

http://tinyurl.com/y937bvp9

So we probably want an ack from the neutron team before approving this.

Patch-set: 5