Document when start_nova_compute does not run multi-cell
The ``local nomulticellflag="$1"`` line in the
start_nova_compute function confused me because nothing
in devstack itself passes a parameter to that function.
It happens in grenade and some other devstack plugins.
This change adds some documentation to the function explaining
when it does not run nova-compute in a multi-cell tiered
deployment.

Change-Id: I28c54624520424e4338edf87c6b556a5c642f758