DNM: Test multi-cell (2)
The dvr-ha multinode job that runs in the devstack
experimental queue is a 3-node job, so it would be
cool to test out real multiple cells. We know things
like migration operations between cells aren't supported
and that might cause some tests to fail, but let's see
what blows up.

Change-Id: Ia39a4d4e4f0e01852ffc4218d74c4dc96607391f