Update patch set 2
Patch Set 2:

So this LGTM, but I'm just curious isn't there a race condition here on a node that doesn't have nova-api? We're calling the nova-api to check the services list, but in a multinode env is there guarantee the subnodes will be starting their compute after the primary node deploys the api server?

Patch-set: 2