Update patch set 7
Patch Set 7: Code-Review-1

We shouldn't do this until after RC1. There is too much risk for breaking a bunch of 3rd party CIs that run on Cinder/Neutron for this to be worthwhile right now. We already lost several hours on the day of Pike feature freeze last week because of the first attempt.

Patch-set: 7