Update patch set 1
Patch Set 1: Code-Review-1

Those steps from the tc resolution are for community support removal and are independent of devstack. The postgres support in devstack is deprecated, mostly unmaintained, untested, and it will be removed. I'm not sure if someone was planning on doing it in pike (it's getting kinda close) but the deprecation statement is still valid. If people want to continue to support postgres with devstack they can write a devstack plugin that sets up postgres after it is removed.

Patch-set: 1
Label: Code-Review=-1