Update patch set 1
Patch Set 1:

When you look at the gate we don't test devstack changes on postgres. Nor do the majority of the projects out there test with it. (neutron apparently being the exception) Also, the amount of people looking at the code is very limited. I can't remember the last time I saw a postgres change or bug filed in devstack. Even if there was one I honestly wouldn't be comfortable reviewing it without any of that testing. If it's not tested it's broken. While you might be using it successfully that doesn't really reflect it's status. It was deprecated for all of these reasons, and it's disingenuous to our users to say it's something we're supporting moving forward.

Patch-set: 1