Don't reinstall python-virtualenv on infra nodes
In the original change I said "for infra nodes, it shouldn't do
anything anyway ...".  Well that was pre-Fedora 26 :)

It may be that the referenced issues in the comment is a bug, or it
may be that it was always intended to work this way and the problem is
ours; dnf went from version 1 -> 2 in Fedora 26 so there's way too
much to find a simple root cause ATM.

As described in the comment, take a simpler approach of skipping this
on infra nodes.  pip-and-virtualenv in dib should have installed the
latest pip, virtualenv and setuptools, so we don't want to fiddle with
that anyway.

[1] https://review.openstack.org/#/c/338998/

Change-Id: Ib300b58377a0d0fe1bd7444c71acdb9a87dc033b